Struct google_licensing1::LicenseAssignmentMethods

pub struct LicenseAssignmentMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on licenseAssignment resources. It is not used directly, but through the Licensing hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_licensing1 as licensing1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use licensing1::Licensing;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Licensing::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `delete(...)`, `get(...)`, `insert(...)`, `list_for_product(...)`, `list_for_product_and_sku(...)`, `patch(...)` and `update(...)`
// to build up your call.
let rb = hub.license_assignments();

Methods

impl<'a, C, A> LicenseAssignmentMethods<'a, C, A>

fn delete(&self, product_id: &str, sku_id: &str, user_id: &str) -> LicenseAssignmentDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Revoke License.

Arguments

productId - Name for product

skuId - Name for sku

userId - email id or unique Id of the user

fn list_for_product(&self, product_id: &str, customer_id: &str) -> LicenseAssignmentListForProductCall<'a, C, A>

Create a builder to help you perform the following task:

List license assignments for given product of the customer.

Arguments

productId - Name for product

customerId - CustomerId represents the customer for whom licenseassignments are queried

fn get(&self, product_id: &str, sku_id: &str, user_id: &str) -> LicenseAssignmentGetCall<'a, C, A>

Create a builder to help you perform the following task:

Get license assignment of a particular product and sku for a user

Arguments

productId - Name for product

skuId - Name for sku

userId - email id or unique Id of the user

fn list_for_product_and_sku(&self, product_id: &str, sku_id: &str, customer_id: &str) -> LicenseAssignmentListForProductAndSkuCall<'a, C, A>

Create a builder to help you perform the following task:

List license assignments for given product and sku of the customer.

Arguments

productId - Name for product

skuId - Name for sku

customerId - CustomerId represents the customer for whom licenseassignments are queried

fn update(&self, request: LicenseAssignment, product_id: &str, sku_id: &str, user_id: &str) -> LicenseAssignmentUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Assign License.

Arguments

request - No description provided.

productId - Name for product

skuId - Name for sku for which license would be revoked

userId - email id or unique Id of the user

fn patch(&self, request: LicenseAssignment, product_id: &str, sku_id: &str, user_id: &str) -> LicenseAssignmentPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Assign License. This method supports patch semantics.

Arguments

request - No description provided.

productId - Name for product

skuId - Name for sku for which license would be revoked

userId - email id or unique Id of the user

fn insert(&self, request: LicenseAssignmentInsert, product_id: &str, sku_id: &str) -> LicenseAssignmentInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Assign License.

Arguments

request - No description provided.

productId - Name for product

skuId - Name for sku