Module google_licensing1::api

Structs

Empty
A generic empty message that you can re-use to avoid defining duplicated empty messages in your APIs. A typical example is to use it as the request or the response type of an API method. For instance: service Foo { rpc Bar(google.protobuf.Empty) returns (google.protobuf.Empty); } The JSON representation for Empty is empty JSON object {}.

LicenseAssignment
Representation of a license assignment.

LicenseAssignmentDeleteCall
Revoke a license.

LicenseAssignmentGetCall
Get a specific user's license by product SKU.

LicenseAssignmentInsert
Representation of a license assignment.

LicenseAssignmentInsertCall
Assign a license.

LicenseAssignmentList
There is no detailed description.

LicenseAssignmentListForProductAndSkuCall
List all users assigned licenses for a specific product SKU.

LicenseAssignmentListForProductCall
List all users assigned licenses for a specific product SKU.

LicenseAssignmentMethods
A builder providing access to all methods supported on licenseAssignment resources. It is not used directly, but through the Licensing hub.

LicenseAssignmentPatchCall
Reassign a user's product SKU with a different SKU in the same product. This method supports patch semantics.

LicenseAssignmentUpdateCall
Reassign a user's product SKU with a different SKU in the same product.

Licensing
Central instance to access all Licensing related resource activities

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.